requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
scikit-learn==1.7.2
joblib==1.5.2
numpy==2.3.4
pillow==12.0.0
python-multipart==0.0.20
torch==2.6.0
torchvision==0.21.0
pandas
matplotlib
seaborn
mlflow
shap
